requirements.txt
# requirements (for Python 3.11 + macOS 12)

# === LangChain + OpenAI (Stable Combo) ===
openai==1.16.1
langchain==0.1.17
langchain-openai==0.0.8
langchain-community==0.0.36

# === Streamlit Chat UI ===
streamlit==1.33.0
streamlit-chat==0.1.1
streamlit-feedback==0.1.2

# === Core ML Stack ===
scikit-learn==1.4.2
imbalanced-learn==0.12.0
tensorflow==2.15.0
keras==2.15.0

# === Fairness + Explainability ===
shap==0.44.0
fairlearn==0.10.0

# === Utility + Parsing ===
python-dotenv==1.0.1
pyparsing==3.1.2
pandas==2.2.1
